chore: refactored to use hono + cloudflare workers
