--- .prettierrc
@@ -1,6 +1,0 @@
-{
-  "printWidth": 90,
-  "singleQuote": true,
-  "semi": true,
-  "useTabs": false
-}
