+++ tsconfig.json
@@ -1,22 +1,23 @@
 {
-  "compilerOptions": {
-    // enable latest features
-    "lib": ["esnext"],
-    "module": "NodeNext",
-    "target": "esnext",
+	"compilerOptions": {
+		"target": "ES2021",
+		"lib": ["ES2021"],
+		"module": "ESNext",
+		"moduleResolution": "bundler",
+		"types": ["@cloudflare/workers-types"],
 
-    // if TS 5.x+
-    "moduleResolution": "NodeNext",
-    "noEmit": true,
-    "moduleDetection": "force",
+		"noEmit": true,
+		"isolatedModules": true,
+		"esModuleInterop": true,
+		"moduleDetection": "force",
+		"resolveJsonModule": true,
+		"allowSyntheticDefaultImports": true,
 
-    "jsx": "react-jsx", // support JSX
-    "allowJs": true, // allow importing `.js` from `.ts`
-    "esModuleInterop": true, // allow default imports for CommonJS modules
-
-    // best practices
-    "strict": true,
-    "forceConsistentCasingInFileNames": true,
-    "skipLibCheck": true
-  }
+		"strict": true,
+		"forceConsistentCasingInFileNames": true,
+		"skipLibCheck": true,
+		"allowJs": true
+	},
+	"include": ["src/**/*"],
+	"exclude": ["node_modules", "api"]
 }
