--- vercel.json
@@ -1,3 +1,0 @@
-{
-  "rewrites": [{ "source": "/:path*", "destination": "/api/:path*" }]
-}
